Used bigdecimal instead of double and added changes accordingly

+++ src/com/bank/models/Account.java
@@ -1,10 +1,12 @@
 package com.bank.models;
 
+import java.math.BigDecimal;
+
 public class Account {
 
     private int id;
     private String accountHolder;
-    private double balance;
+    private BigDecimal balance;
 
     /**
      * Constructor to initialize an account based on the name and starting amount
@@ -12,7 +14,7 @@ public class Account {
      * @param accountHolder Name of the account holder
      * @param balance Amount to be added to the balance
      */
-    public Account (final String accountHolder, final double balance) {
+    public Account (final String accountHolder, final BigDecimal balance) {
         this.accountHolder = accountHolder;
         this.balance = balance;
     }
@@ -24,7 +26,7 @@ public class Account {
      * @param balance Amount to be added to the balance
      * @param id id of the account holder
      */
-    public Account (final String accountHolder, final double balance, final int id) {
+    public Account (final String accountHolder, final BigDecimal balance, final int id) {
         this.accountHolder = accountHolder;
         this.balance = balance;
         this.id = id;
@@ -50,7 +52,7 @@ public class Account {
      * Fetches the balance from account object
      * @return Returns the balance in the account.
      */
-    public double getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 }
