+++ src/com/bank/models/Transaction.java
@@ -1,10 +1,11 @@
 package com.bank.models;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public record Transaction(double amount, String transactionType, LocalDateTime timeStamp) {
+public record Transaction(BigDecimal amount, String transactionType, LocalDateTime timeStamp) {
 
-    public static Transaction getTransaction(final double amount, final String transactionType) {
+    public static Transaction getTransaction(final BigDecimal amount, final String transactionType) {
         return new Transaction(amount, transactionType, LocalDateTime.now());
     }
